authentication and authorization established using jwt

+++ controllers/api/v1/posts_api.js
@@ -21,17 +21,21 @@ module.exports.destroy = async function(req, res){
     
     try{
         let post = await Post.findById(req.params.id);
-        
-            post.remove();
-            await Comment.deleteMany({post: req.params.id});
+            if(post.user == req.user.id){
+                post.remove();
+                await Comment.deleteMany({post: req.params.id});
 
-            return res.json(200, {
-                message: 'Post and Associated Comments deleted'                
-            });
-            
+                return res.json(200, {
+                    message: 'Post and Associated Comments deleted'                
+                });
+            }else{
+                return res.json(402, {
+                    message: 'You cannot delete this Post!'
+                });
+            }
 
     }catch(err){
-        
+        console.log('********', err);
         return res.json(500, {
             message: 'Internal Server Error'
         });
